requirements.txt
asgiref==3.8.1
Django==5.1.4
future==1.0.0
gunicorn==23.0.0
Pillow==11.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
sqlparse==0.5.3
tzdata==2024.2
